applets/notes: Also save/restore cursor position in TextArea
Very straightforward addition to the scrolling management code. Can be
tested by assigning a global shortcut to an applet, so that it can
become focused without actually clicking on it with a mouse and messing
with cursor position. Text selection is not handled, and I don't think
we should even bother with it.